+++ include/pacbio/ccs/SparseAlignment.h
@@ -50,117 +50,74 @@
 #include <seqan/seeds.h>
 #include <seqan/sequence.h>
 
+#include <pacbio/align/SparseAlignment.h>
+
 #include <pacbio/ccs/ChainSeeds.h>
 #include <pacbio/exception/CCSExceptions.h>
 
+#include <pbcopper/align/Seed.h>
+#include <pbcopper/qgram/Index.h>
+
 namespace PacBio {
 namespace CCS {
-namespace {
 
-inline size_t SafeSubtract(size_t size, size_t k) { return size > k ? size - k : 0; }
-template <typename TShape>
-class HpHasher
+///
+/// Find all matching seeds between two DNA sequences
+///
+/// \sa PacBio::Align::FindSeeds
+///
+/// For PacBio::CCS, homopolymer filtering is always enabled.
+///
+/// \param[in]  seq1    The first, or query, sequence
+/// \param[in]  seq2    The second, or reference, sequence
+///
+/// \return Seeds collection containing all hits
+///
+inline PacBio::Align::Seeds FindSeeds(const size_t qGramSize,
+                                      const std::string& seq1,
+                                      const std::string& seq2)
 {
-public:
-    HpHasher(TShape& shape)
-    {
-        using namespace seqan;
-        using std::string;
-
-        const char dna[4] = {'A', 'C', 'G', 'T'};
-
-        for (size_t i = 0; i < 4; i++) {
-            DnaString s = string(length(shape), dna[i]);
-            hashes[i] = hash(shape, begin(s));
-        }
-    }
-
-    inline bool operator()(const unsigned h) const
-    {
-        if (h == hashes[0] || h == hashes[1] || h == hashes[2] || h == hashes[3]) return true;
-
-        return false;
-    }
-
-private:
-    unsigned hashes[4];
-};
-
-}  // anonymous namespace
-
-template <typename TConfig>
-void FindSeeds(seqan::SeedSet<seqan::Seed<seqan::Simple>>& seeds, const seqan::DnaString& seq1,
-               const seqan::DnaString& seq2)
-{
-    using namespace seqan;
-
-    typedef Shape<Dna, typename TConfig::ShapeType> TShape;
-
-    Index<DnaString, typename TConfig::IndexType> index(seq1);
-    TShape shape = indexShape(index);
-    HpHasher<TShape> isHomopolymer(shape);
-    auto start = begin(seq2);
-    size_t end = SafeSubtract(length(seq2) + 1, TConfig::Size);
-
-    hashInit(shape, start);
-
-    for (size_t j = 0; j < end; j++) {
-        if (isHomopolymer(hashNext(shape, start + j))) continue;
-
-        auto hits = getOccurrences(index, shape);
-
-        for (const auto& i : hits) {
-            Seed<Simple> seed(i, j, TConfig::Size);
-
-#ifdef MERGESEEDS
-            if (!addSeed(seeds, seed, 0, Merge()))
-#endif
-            {
-                addSeed(seeds, seed, Single());
-            }
-        }
-    }
+    return PacBio::Align::FindSeeds(qGramSize, seq1, seq2, true);
 }
 
-template <typename TConfig>
-void FindSeeds(
-    std::map<size_t, seqan::SeedSet<seqan::Seed<seqan::Simple>>>& seeds,
-    const seqan::Index<seqan::StringSet<seqan::DnaString>, typename TConfig::IndexType>& index,
-    const seqan::DnaString& seq, const boost::optional<size_t> qIdx = boost::none)
+///
+/// \brief FindSeeds
+///
+/// \sa PacBio::Align::FindSeeds
+///
+/// For PacBio::CCS, homopolymer filtering is always enabled.
+///
+/// \param[in]  index   The hashed index on the reference sequence(s)
+/// \param[in]  seq     The query sequence
+/// \param[in]  qIdx    (optional) The index of the query sequence, so it can be ignored
+///
+/// \return map containing Seeds for each referenceIndex with a hit
+///
+inline std::map<size_t, PacBio::Align::Seeds> FindSeeds(
+        const PacBio::QGram::Index& index,
+        const std::string& seq,
+        const boost::optional<size_t> qIdx)
 {
-    using namespace seqan;
-    using namespace std;
+    return PacBio::Align::FindSeeds(index, seq, qIdx, true);
+}
 
-    typedef Shape<Dna, typename TConfig::ShapeType> TShape;
-
-    TShape shape = indexShape(index);
-    HpHasher<TShape> isHomopolymer(shape);
-    auto start = begin(seq);
-    size_t end = SafeSubtract(length(seq) + 1, TConfig::Size);
-
-    hashInit(shape, start);
-
-    for (size_t i = 0; i < end; i++) {
-        if (isHomopolymer(hashNext(shape, start + i))) continue;
-
-        auto hits = getOccurrences(index, shape);
-
-        for (const auto& hit : hits) {
-            size_t rIdx = getValueI1(hit);
-
-            if (qIdx && rIdx == *qIdx) continue;
-
-            size_t j = getValueI2(hit);
-            Seed<Simple> seed(i, j, TConfig::Size);
-
-#ifdef MERGESEEDS
-            if (!addSeed(seeds[rIdx], seed, 0, Merge()))
-#endif
-            {
-                addSeed(seeds[rIdx], seed, Single());
-            }
-        }
-    }
+///
+/// \brief FindSeeds
+///
+/// \sa PacBio::Align::FindSeeds
+///
+/// For PacBio::CCS, homopolymer filtering is always enabled.
+///
+/// \param[in]  index   The hashed index on the reference sequence(s)
+/// \param[in]  seq     The query sequence
+///
+/// \return map containing Seeds for each referenceIndex with a hit
+///
+inline std::map<size_t, PacBio::Align::Seeds> FindSeeds(
+        const PacBio::QGram::Index& index,
+        const std::string& seq)
+{
+    return PacBio::Align::FindSeeds(index, seq, boost::none, true);
 }
 
 template <size_t TSize, typename TContainer>
@@ -244,38 +201,47 @@ struct FindSeedsConfig
     static const size_t Size = TSize;
 };
 
-template <size_t TSize>
-seqan::String<seqan::Seed<seqan::Simple>> SparseAlign(const seqan::DnaString& seq1,
-                                                      const seqan::DnaString& seq2)
+///
+/// Generate an SDP alignment from two sequences
+///
+/// \param[in]  qGramSize   qgram size to use for index hashing
+/// \param[in]  seq1        The first, or query, sequence
+/// \param[in]  seq2        The second, or reference, sequence
+///
+/// \return The SDP alignment as a vector of Seeds
+///
+inline std::vector<PacBio::Align::Seed> SparseAlignSeeds(const size_t qGramSize,
+                                                         const std::string& seq1,
+                                                         const std::string& seq2)
 {
-    using namespace seqan;
-
-    typedef FindSeedsConfig<TSize> TConfig;
-
-    SeedSet<Seed<Simple>> seeds;
-    FindSeeds<TConfig>(seeds, seq1, seq2);
-
-    String<Seed<Simple>> chain;
-    ChainSeeds(chain, seeds);
-
-    return chain;
+    const auto config = PacBio::Align::ChainSeedsConfig{ UINT_MAX, 0, 3, 1, -1, -1, INT_MAX };
+    const auto seeds  = PacBio::Align::FindSeeds(qGramSize, seq1, seq2, true);
+    const auto chains = PacBio::Align::ChainSeeds(seeds, config);
+    if (chains.empty())
+        return std::vector<PacBio::Align::Seed>{};
+    return chains[0];
 }
 
-template <size_t TSize>
-std::vector<std::pair<size_t, size_t>> SparseAlign(const std::string& seq1, const std::string& seq2)
+///
+/// Generate an SDP alignment from two sequences
+///
+/// \param[in]  qGramSize   qgram size to use for index hashing
+/// \param[in]  seq1        The first, or query, sequence
+/// \param[in]  seq2        The second, or reference, sequence
+/// \param seq2
+///
+/// \return A vector of pairs, representing Kmer start positions
+///             that match in the query and reference sequences
+///
+inline std::vector<std::pair<size_t, size_t>> SparseAlign(const size_t qGramSize,
+                                                          const std::string& seq1,
+                                                          const std::string& seq2)
 {
-    using namespace seqan;
-    using namespace std;
 
-    DnaString a = seq1, b = seq2;
-    auto chain = SparseAlign<TSize>(a, b);
-
-    vector<pair<size_t, size_t>> result;
-
-    for (const auto& s : chain) {
-        result.push_back(make_pair(beginPositionH(s), beginPositionV(s)));
-    }
-
+    std::vector<std::pair<size_t, size_t>> result;
+    const auto chain = SparseAlignSeeds(qGramSize, seq1, seq2);
+    for (const auto& s : chain)
+        result.emplace_back(s.BeginPositionH(), s.BeginPositionV());
     return result;
 }