+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add Fisher's exact test
  - Juliet JSON and HTMl output for HIV-1 pol region
  - Call codon-wise
+ - Use pbcopper's q-gram index for sparse alignment
 
 ## [2.0.4]